Changes related to PR 10604

+++ contracts/extensions/CoinMachine.sol
@@ -89,7 +89,7 @@ contract CoinMachine is ColonyExtension {
 
   /// @notice Called when uninstalling the extension
   function uninstall() public override auth {
-    selfdestruct(payable(uint160(address(colony))));
+    selfdestruct(payable(address(colony)));
   }
 
   /// @notice Must be called before any sales can be made
