+++ contracts/extensions/FundingQueue.sol
@@ -108,7 +108,7 @@ contract FundingQueue is ColonyExtension, PatriciaTreeProofs {
 
   /// @notice Called when uninstalling the extension
   function uninstall() public override auth {
-    selfdestruct(payable(uint160(address(colony))));
+    selfdestruct(payable(address(colony)));
   }
 
   function createProposal(
