+++ contracts/extensions/TokenSupplier.sol
@@ -92,7 +92,7 @@ contract TokenSupplier is ColonyExtension {
 
   /// @notice Called when uninstalling the extension
   function uninstall() public override auth {
-    selfdestruct(payable(uint160(address(colony))));
+    selfdestruct(payable(address(colony)));
   }
 
   /// @notice Initialise the extension, must be called before any tokens can be issued
