+++ contracts/extensions/VotingReputation.sol
@@ -179,7 +179,7 @@ contract VotingReputation is ColonyExtension, PatriciaTreeProofs {
 
   /// @notice Called when uninstalling the extension
   function uninstall() public override auth {
-    selfdestruct(payable(uint160(address(colony))));
+    selfdestruct(payable(address(colony)));
   }
 
   // Data structures
